feat: add QuantitySelector component
