+++ README.md
@@ -2,8 +2,8 @@
 ## 📡 1단계 - 상품 상세 페이지 & 상품 결제하기 Form 구현하기
 ### ✅ 기능 목록
 - [x] chakra-ui 설치
-- [] 상품 상세페이지 UI 구현
-- [] 강품 결제페이지 UI 구현
+- [x] 상품 상세페이지 UI 구현
+- [] 상품 결제페이지 UI 구현
 - [] `/api/v1/products/{productId}/detail`, `/api/v1/products/{productId}/options`에 대한 API 구현
 - [] 없는 상품의 경우 메인 페이지로 리다이렉트
 - [] 나에게 선물하기 버튼 클릭 시 로그인이 되어있지 않다면 로그인 페이지로 리다이렉트
